+++ src/components/generator/dnd/Generator.tsx
@@ -1,28 +1,54 @@
+import '@atlaskit/css-reset';
+import { DragDropContext, Droppable } from 'react-beautiful-dnd';
+import { v4 as uuid } from 'uuid';
 import styled from '@emotion/styled';
-import DndContainer from 'components/generator/dnd/DndContainer';
-import React, { useEffect, useState } from 'react';
+import { useRecoilState } from 'recoil';
+import { componentsState } from 'atoms/components';
+import DndComponent from './DndComponent';
 
 const Generator = () => {
-  const [start, setStart] = useState(false);
+  const [components, setComponents] = useRecoilState(componentsState);
 
-  useEffect(() => {
-    setStart(true);
-  }, []);
+  const handleChange = (result: any) => {
+    if (!result.destination) return;
+    const items = [...components];
+    const [reorderedItem] = items.splice(result.source.index, 1);
+    items.splice(result.destination.index, 0, reorderedItem);
+    setComponents(items);
+  };
 
-  return <Container>{start ? <DndContainer /> : null}</Container>;
+  const addTextEditor = () => {
+    setComponents((oldComponents: any[]) => [
+      ...oldComponents,
+      {
+        id: `${uuid()}`,
+        code: '',
+        type: 'text',
+      },
+    ]);
+  };
+
+  return (
+    <DragDropContext onDragEnd={handleChange}>
+      <Droppable droppableId="cardlists">
+        {provided => (
+          <div className="cardlists" {...provided.droppableProps} ref={provided.innerRef}>
+            {components.map((component: any, i: number) => (
+              <DndComponent key={`test-${component.id}`} component={component} componentIndex={i} />
+            ))}
+            {provided.placeholder}
+          </div>
+        )}
+      </Droppable>
+      <BtnAdd onClick={addTextEditor}>추가</BtnAdd>
+    </DragDropContext>
+  );
 };
 
 export default Generator;
 
-const Container = styled.div`
-  background-color: #171b21;
-  padding: 70px 20px 70px;
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  height: 100%;
-
-  h3 {
-    color: white;
-  }
+const BtnAdd = styled.button`
+  color: white;
+  font-weight: 700;
+  font-size: 20px;
 `;
